code/requirements.txt
contourpy==1.3.3
cycler==0.12.1
dill==0.4.0
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
psutil==7.1.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
qiskit==2.2.2
qiskit-aer==0.17.2
rustworkx==0.17.1
scipy==1.16.2
six==1.17.0
stevedore==5.5.0
typing_extensions==4.15.0
